ENG-4068 fix export issue where uso is passed around as an int parameter instead of a long

+++ src/frontend/org/voltdb/ExecutionSite.java
@@ -2914,7 +2914,7 @@ implements Runnable, SiteTransactionConnection, SiteProcedureConnection, SiteSna
 
     @Override
     public void exportAction(boolean syncAction,
-                             int ackOffset,
+                             long ackOffset,
                              Long sequenceNumber,
                              Integer partitionId,
                              String tableSignature)
